chore: add engines to package.json
