+++ .github/workflows/run-tests.yml
@@ -20,6 +20,7 @@ jobs:
           - windows-latest
           - macos-latest
         node:
+          - 22.12.0
           - 22
           - 23
           - 24
