+++ scripts/generate-docs.ts
@@ -8,7 +8,7 @@ import {Client} from '@modelcontextprotocol/sdk/client/index.js';
 import {StdioClientTransport} from '@modelcontextprotocol/sdk/client/stdio.js';
 import type {Tool} from '@modelcontextprotocol/sdk/types.js';
 import {ToolCategories} from '../build/src/tools/categories.js';
-import {cliOptions} from '../build/src/index.js';
+import {cliOptions} from '../build/src/main.js';
 import fs from 'fs';
 
 const MCP_SERVER_PATH = 'build/src/index.js';
